backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.1
pydantic-settings==2.1.0
langchain==0.0.350
langchain-community==0.0.10
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
together==0.2.10
llama-parse==0.1.4
aiofiles==23.2.1
pythai==1.0.0
numpy==1.24.4
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
cryptography==41.0.8
bcrypt==4.1.2
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1